# perf : 복합 인덱스를 이용한 조회 성능 향상 시도