--- a/src/main/java/org/jungppo/bambooforest/chatbot/domain/entity/ChatBotPurchaseEntity.java
+++ b/src/main/java/org/jungppo/bambooforest/chatbot/domain/entity/ChatBotPurchaseEntity.java
@@ -8,6 +8,7 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.Index;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
@@ -23,7 +24,10 @@ import org.jungppo.bambooforest.member.domain.entity.MemberEntity;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "chatbot_purchase")
+@Table(name = "chatbot_purchase",
+        indexes = {
+                @Index(name = "idx_member_id_created_at", columnList = "member_id, created_at")
+        })
 public class ChatBotPurchaseEntity extends JpaBaseEntity {
 
     @Id