--- a/src/main/java/org/jungppo/bambooforest/payment/domain/entity/PaymentEntity.java
+++ b/src/main/java/org/jungppo/bambooforest/payment/domain/entity/PaymentEntity.java
@@ -8,6 +8,7 @@ import jakarta.persistence.FetchType;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
+import jakarta.persistence.Index;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
@@ -25,7 +26,10 @@ import org.jungppo.bambooforest.member.domain.entity.MemberEntity;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "payment")
+@Table(name = "payment",
+        indexes = {
+                @Index(name = "idx_member_id_status_created_at", columnList = "member_id, status, created_at")
+        })
 public class PaymentEntity extends JpaBaseEntity {
 
     @Id