<!--
PR 이름 컨벤션
feat: ~~(#issueNum)
-->

## 📌 관련 이슈

- #17, #26

## ✨ PR 세부 내용
테스트 환경은 해커톤 환경에 기반하여 2000명의 사용자로 예상하여 구성하였습니다.

1. **payment entity에 복합 인덱스를 이용한 조회 성능 향상 시도**
   - 2000명의 인원이 50번의 결제를 시도하였을 경우
     - EXPLAIN ANALYZE
       -  인덱스 적용 전 
         ![image](https://github.com/user-attachments/assets/c1933a7f-fd5d-429e-8891-72deb21003d6)
       -  인덱스 적용 후
         ![image](https://github.com/user-attachments/assets/2236af44-b0d0-4b7e-900c-27cacb5a02ba)
     - 실제 실행
       -  인덱스 적용 전 
           - 평균 실행 시간: 511,865,800 ns
       -  인덱스 적용 후
           - 평균 실행 시간: 478,125,700 ns
        
     - **분석 결과**
       - Execution Time이 약 15.494 ms에서 0.619 ms로 감소하여 실제 쿼리 실행 시간이 단축되었습니다.
       - 인덱스 적용 후 평균 실행 시간이 약 33,740,100 ns 감소하여 약 6.59%의 성능 향상이 있었습니다.

2. **chatBotPurchase entity에 복합 인덱스를 이용한 조회 성능 향상 시도**
   - 2000명의 인원이 2개의 챗봇을 구매하였을 경우
     - EXPLAIN ANALYZE
       -  인덱스 적용 전 
         ![image](https://github.com/user-attachments/assets/7f4117d8-c949-49f7-9ca7-449252534390)
       -  인덱스 적용 후
         ![image](https://github.com/user-attachments/assets/b22cd7f9-6f58-474f-b123-20f0e476ad2c)
     - 실제 실행
       -  인덱스 적용 전 
          - 평균 실행 시간: 489,639,500 ns
       -  인덱스 적용 후
           - 평균 실행 시간: 433,720,750 ns
        
     - **분석 결과**
       - Execution Time이 약 0.428 ms에서 0.170 ms로 감소하여 실제 쿼리 실행 시간이 단축되었습니다.
       - 인덱스 적용 후 평균 실행 시간이 약 55,918,750 ns 감소하여 약 11.42%의 성능 향상이 있었습니다.