@DongHyeonka 2024-07-31
복합 인덱스를 사용하여 조회 성능 향상 하신 내용 잘봤습니다! 다음에 저도 성능 향상 시킬 경우가 생기면 고려해보겠습니다!